feat: implement retry on init failure #31222
- refactor and split BatchOperationExecutionScheduler into two separate classes
- add retry-loop in scheduler in case an error happened
