+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/batchoperation/BatchOperationExecutionScheduler.java
@@ -7,42 +7,29 @@
  */
 package io.camunda.zeebe.engine.processing.batchoperation;
 
-import com.google.common.collect.Lists;
+import com.google.common.base.Strings;
+import com.google.common.math.IntMath;
+import io.camunda.search.exception.CamundaSearchException;
+import io.camunda.search.exception.CamundaSearchException.Reason;
 import io.camunda.zeebe.engine.EngineConfiguration;
 import io.camunda.zeebe.engine.metrics.BatchOperationMetrics;
-import io.camunda.zeebe.engine.processing.batchoperation.itemprovider.ItemProvider.Item;
-import io.camunda.zeebe.engine.processing.batchoperation.itemprovider.ItemProvider.ItemPage;
 import io.camunda.zeebe.engine.processing.batchoperation.itemprovider.ItemProviderFactory;
 import io.camunda.zeebe.engine.state.batchoperation.PersistedBatchOperation;
 import io.camunda.zeebe.engine.state.immutable.BatchOperationState;
 import io.camunda.zeebe.engine.state.immutable.ScheduledTaskState;
-import io.camunda.zeebe.msgpack.value.StringValue;
-import io.camunda.zeebe.protocol.impl.record.UnifiedRecordValue;
-import io.camunda.zeebe.protocol.impl.record.value.batchoperation.BatchOperationChunkRecord;
-import io.camunda.zeebe.protocol.impl.record.value.batchoperation.BatchOperationError;
-import io.camunda.zeebe.protocol.impl.record.value.batchoperation.BatchOperationExecutionRecord;
-import io.camunda.zeebe.protocol.impl.record.value.batchoperation.BatchOperationInitializationRecord;
-import io.camunda.zeebe.protocol.impl.record.value.batchoperation.BatchOperationItem;
-import io.camunda.zeebe.protocol.impl.record.value.batchoperation.BatchOperationPartitionLifecycleRecord;
-import io.camunda.zeebe.protocol.record.intent.BatchOperationChunkIntent;
-import io.camunda.zeebe.protocol.record.intent.BatchOperationExecutionIntent;
-import io.camunda.zeebe.protocol.record.intent.BatchOperationIntent;
 import io.camunda.zeebe.protocol.record.value.BatchOperationErrorType;
-import io.camunda.zeebe.stream.api.FollowUpCommandMetadata;
 import io.camunda.zeebe.stream.api.ReadonlyStreamProcessorContext;
 import io.camunda.zeebe.stream.api.StreamProcessorLifecycleAware;
 import io.camunda.zeebe.stream.api.scheduling.AsyncTaskGroup;
 import io.camunda.zeebe.stream.api.scheduling.TaskResult;
 import io.camunda.zeebe.stream.api.scheduling.TaskResultBuilder;
 import java.time.Duration;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
-import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -60,25 +47,18 @@ public class BatchOperationExecutionScheduler implements StreamProcessorLifecycl
   public static final String ERROR_MSG_FAILED_FIRST_CHUNK_APPEND =
       "Unable to append first chunk of batch operation items. Number of items: %d";
   private static final Logger LOG = LoggerFactory.getLogger(BatchOperationExecutionScheduler.class);
-
-  private static final UnifiedRecordValue EMPTY_EXECUTION_RECORD =
-      new BatchOperationExecutionRecord().setBatchOperationKey(-1L);
-  private static final UnifiedRecordValue EMPTY_INITIALIZATION_RECORD =
-      new BatchOperationInitializationRecord()
-          .setBatchOperationKey(-1L)
-          .setSearchQueryPageSize(0) // random int
-          // random cursor string. 1024 chars should be enough for any cursor without sorting
-          .setSearchResultCursor(RandomStringUtils.insecure().next(1024));
-
-  private final Duration pollingInterval;
-  private final int chunkSize;
-  private final int queryPageSize;
+  private static final Set<Reason> FAIL_IMMEDIATELY_REASONS =
+      Set.of(
+          Reason.NOT_FOUND, Reason.NOT_UNIQUE, Reason.SECONDARY_STORAGE_NOT_SET, Reason.FORBIDDEN);
+  private final Duration initialPollingInterval;
+  private final Duration initialRetryDelay;
+  private final Duration maxRetryDelay;
+  private final int maxRetries;
+  private final int backoffFactor;
 
   private final BatchOperationState batchOperationState;
   private ReadonlyStreamProcessorContext processingContext;
-  private final ItemProviderFactory itemProviderFactory;
-  private final BatchOperationMetrics metrics;
-  private final int partitionId;
+  private final BatchOperationInitializer batchOperationInitializer;
 
   /** Marks if this scheduler is currently executing or not. */
   private final AtomicBoolean executing = new AtomicBoolean(false);
@@ -92,31 +72,33 @@ public class BatchOperationExecutionScheduler implements StreamProcessorLifecycl
       final int partitionId,
       final BatchOperationMetrics metrics) {
     batchOperationState = scheduledTaskStateFactory.get().getBatchOperationState();
-    this.itemProviderFactory = itemProviderFactory;
-    pollingInterval = engineConfiguration.getBatchOperationSchedulerInterval();
-    chunkSize = engineConfiguration.getBatchOperationChunkSize();
-    queryPageSize = engineConfiguration.getBatchOperationQueryPageSize();
-    this.metrics = metrics;
-    this.partitionId = partitionId;
+    initialPollingInterval = engineConfiguration.getBatchOperationSchedulerInterval();
+    initialRetryDelay = engineConfiguration.getBatchOperationQueryRetryInitialDelay();
+    maxRetryDelay = engineConfiguration.getBatchOperationQueryRetryMaxDelay();
+    maxRetries = engineConfiguration.getBatchOperationQueryRetryMax();
+    backoffFactor = engineConfiguration.getBatchOperationQueryRetryBackoffFactor();
+    batchOperationInitializer =
+        new BatchOperationInitializer(
+            itemProviderFactory, engineConfiguration, partitionId, metrics);
   }
 
   @Override
   public void onRecovered(final ReadonlyStreamProcessorContext context) {
     processingContext = context;
-    scheduleExecution();
+    scheduleExecution(initialPollingInterval);
   }
 
   @Override
   public void onResumed() {
-    scheduleExecution();
+    scheduleExecution(initialPollingInterval);
   }
 
   /** Schedules the next execution of the batch operation scheduler run. */
-  private void scheduleExecution() {
+  private void scheduleExecution(final Duration nextDelay) {
     if (!executing.get()) {
       processingContext
           .getScheduleService()
-          .runDelayedAsync(pollingInterval, this::execute, AsyncTaskGroup.BATCH_OPERATIONS);
+          .runDelayedAsync(nextDelay, this::execute, AsyncTaskGroup.BATCH_OPERATIONS);
     } else {
       LOG.warn("Execution is already in progress, skipping scheduling.");
     }
@@ -130,76 +112,69 @@ public class BatchOperationExecutionScheduler implements StreamProcessorLifecycl
    * @return the task result containing the commands to be executed
    */
   private TaskResult execute(final TaskResultBuilder taskResultBuilder) {
+    var nextDelay = initialPollingInterval;
     try {
       LOG.trace("Looking for the next pending batch operation to execute (scheduled).");
       executing.set(true);
-      batchOperationState
-          .getNextPendingBatchOperation()
-          .ifPresent(bo -> initializeBatchOperation(bo, taskResultBuilder));
-      return taskResultBuilder.build();
+      nextDelay =
+          executeRetrying(batchOperationState.getNextPendingBatchOperation(), taskResultBuilder);
     } finally {
       executing.set(false);
-      scheduleExecution();
+      scheduleExecution(nextDelay);
     }
+    return taskResultBuilder.build();
   }
 
-  private void initializeBatchOperation(
-      final PersistedBatchOperation batchOperation, final TaskResultBuilder taskResultBuilder) {
-    if (batchOperation.isSuspended()) {
-      LOG.trace("Batch operation {} is suspended.", batchOperation.getKey());
-      return;
+  private Duration executeRetrying(
+      final Optional<PersistedBatchOperation> batchOperation,
+      final TaskResultBuilder taskResultBuilder) {
+    if (batchOperation.isEmpty()) {
+      return initialPollingInterval;
     }
 
-    if (!validateNoReInitialization(batchOperation)) {
-      return;
+    if (!validateNoReInitialization(batchOperation.get())) {
+      return initialPollingInterval;
     }
 
-    final var itemProvider = itemProviderFactory.fromBatchOperation(batchOperation);
-
-    // use overall state variable for all parameters that are used in the loop
-    final InitLoopState loopState = new InitLoopState(batchOperation);
-    while (loopState.hasNextPage()) {
-      try {
-        loopState.page =
-            itemProvider.fetchItemPage(
-                loopState.lastSearchResultCursor, loopState.searchResultPageSize);
-      } catch (final Exception e) {
-        LOG.error(
-            "Failed to query keys for batch operation with key {}. It will be removed from queue",
-            batchOperation.getKey(),
-            e);
-        appendFailedCommand(
+    final var result =
+        batchOperationInitializer.initializeBatchOperation(batchOperation.get(), taskResultBuilder);
+    if (!result.isSuccess()) {
+      if (shouldFailImmediately(result.exception())
+          || initializing.get().numAttempts >= maxRetries) {
+        batchOperationInitializer.appendFailedCommand(
             taskResultBuilder,
-            batchOperation,
-            ExceptionUtils.getStackTrace(e),
+            batchOperation.get(),
+            ExceptionUtils.getStackTrace(result.exception()),
             BatchOperationErrorType.QUERY_FAILED);
-        return;
+        return initialPollingInterval;
       }
-
-      // Then try to append the items to the batch operation
-      final boolean appendedChunks =
-          appendChunks(batchOperation, taskResultBuilder, loopState.page.items());
-      if (appendedChunks) {
-        // everything went normally, so we can continue with the next page in the next loop
-
-        loopState.chunksAppendedThisRun = true;
-        loopState.lastSearchResultCursor = loopState.page.endCursor();
-        loopState.keysAdded += loopState.page.items().size();
-        if (loopState.page.isLastPage()) {
-          // If we have reached the last page, we can finalize the initialization and start the BO
-          // we always append the EXECUTE command at the end, even if no items were found, so we can
-          // leave the completion logic in that processor.
-          finishInitialization(loopState.batchOperation, taskResultBuilder);
-          startExecutionPhase(taskResultBuilder, loopState);
-
-          // this is the end of the initialization run, so we can break out of the loop
-          return;
-        }
-      } else {
-        handleFailedChunkAppend(taskResultBuilder, loopState);
-        return;
-      }
+      final var calculatedRetryDelay =
+          initialRetryDelay.multipliedBy(
+              IntMath.pow(backoffFactor, initializing.get().numAttempts()));
+      final var nextRetryDelay =
+          maxRetryDelay.compareTo(calculatedRetryDelay) < 0 ? maxRetryDelay : calculatedRetryDelay;
+      LOG.warn(
+          "Retryable operation failed, retries left: {}, retrying in {} ms. Error: {}",
+          initializing.get().numAttempts(),
+          nextRetryDelay,
+          result.exception().getLocalizedMessage());
+      initializing.set(
+          new ExecutionLoopState(
+              initializing.get().batchOperationKey,
+              Strings.nullToEmpty(initializing.get().searchResultCursor),
+              initializing.get().numAttempts() + 1));
+      return nextRetryDelay;
+    } else {
+      initializing.set(
+          new ExecutionLoopState(result.batchOperationKey(), result.searchResultCursor(), 0));
     }
+
+    return initialPollingInterval;
+  }
+
+  private boolean shouldFailImmediately(final Exception exception) {
+    return exception instanceof CamundaSearchException
+        && FAIL_IMMEDIATELY_REASONS.contains(((CamundaSearchException) exception).getReason());
   }
 
   private boolean validateNoReInitialization(final PersistedBatchOperation batchOperation) {
@@ -207,227 +182,23 @@ public class BatchOperationExecutionScheduler implements StreamProcessorLifecycl
     if (initializingBO != null
         && initializingBO.batchOperationKey == batchOperation.getKey()
         && !Objects.equals(
-            initializingBO.searchResultCursor, batchOperation.getInitializationSearchCursor())) {
+            initializingBO.searchResultCursor, batchOperation.getInitializationSearchCursor())
+        && initializingBO.numAttempts == 0) {
       // If the batch operation is already being initialized, we do not re-initialize it.
       LOG.trace(
           "Batch operation {} is already being executed, skipping re-initialization.",
           batchOperation.getKey());
       return false;
+    } else if (initializingBO == null) {
+      initializing.set(new ExecutionLoopState(batchOperation));
     }
-    initializing.set(
-        new ExecutionLoopState(
-            batchOperation.getKey(), batchOperation.getInitializationSearchCursor(), 0));
 
     return true;
   }
 
-  private void startExecutionPhase(
-      final TaskResultBuilder resultBuilder, final InitLoopState state) {
-    appendExecution(state.batchOperation.getKey(), resultBuilder);
-
-    // start some metrics for the execution phase
-    metrics.recordItemsPerPartition(
-        state.batchOperation.getNumTotalItems() + state.keysAdded,
-        state.batchOperation.getBatchOperationType());
-    metrics.startStartExecuteLatencyMeasure(
-        state.batchOperation.getKey(), state.batchOperation.getBatchOperationType());
-    metrics.startTotalExecutionLatencyMeasure(
-        state.batchOperation.getKey(), state.batchOperation.getBatchOperationType());
-  }
-
-  private boolean finishInitialization(
-      final PersistedBatchOperation batchOperation, final TaskResultBuilder resultBuilder) {
-    final long batchOperationKey = batchOperation.getKey();
-    final var command =
-        new BatchOperationInitializationRecord().setBatchOperationKey(batchOperationKey);
-    LOG.trace("Appending batch operation {} initializing finished command", batchOperationKey);
-
-    final boolean appended =
-        resultBuilder.appendCommandRecord(
-            batchOperationKey,
-            BatchOperationIntent.FINISH_INITIALIZATION,
-            command,
-            FollowUpCommandMetadata.of(b -> b.batchOperationReference(batchOperationKey)));
-    if (appended) {
-      initializing.set(new ExecutionLoopState(batchOperation.getKey(), "finished", 0));
-      metrics.recordInitialized(batchOperation.getBatchOperationType());
-    }
-    return appended;
-  }
-
-  private void handleFailedChunkAppend(
-      final TaskResultBuilder taskResultBuilder, final InitLoopState state) {
-    if (!state.chunksAppendedThisRun) {
-      // the first chunk of this init-run could not be appended. We try to reduce the page size
-      // and retry again. If the pageSize can be halved, we will fail this partition.
-      if (state.searchResultPageSize > 1) {
-        state.searchResultPageSize = state.searchResultPageSize / 2;
-        continueInitialization(taskResultBuilder, state);
-      } else {
-        // If we failed to append the first chunk even when the pageSize is 1,
-        // we need to fail the batch operation. Otherwise, we would be stuck in an infinite loop
-        appendFailedCommand(
-            taskResultBuilder,
-            state.batchOperation,
-            String.format(ERROR_MSG_FAILED_FIRST_CHUNK_APPEND, state.page.items().size()),
-            BatchOperationErrorType.RESULT_BUFFER_SIZE_EXCEEDED);
-      }
-    } else {
-      // The RecordBatch is full, so we need another init run to continue
-      continueInitialization(taskResultBuilder, state);
-    }
-  }
-
-  /**
-   * THis method will create chunk records from the given items and append them to the
-   * taskResultBuilder. This method is atomic. Either all chunks are appended or none.
-   *
-   * @param batchOperation the batch operation to which the chunks belong
-   * @param taskResultBuilder the task result builder to append the chunks to
-   * @param items the items to be chunked and appended
-   * @return true if the chunks were successfully appended, false otherwise
-   */
-  private boolean appendChunks(
-      final PersistedBatchOperation batchOperation,
-      final TaskResultBuilder taskResultBuilder,
-      final List<Item> items) {
-
-    final var chunkRecords =
-        Lists.partition(items, chunkSize).stream()
-            .map(chunkItems -> createChunkRecord(batchOperation, chunkItems))
-            .toList();
-
-    final FollowUpCommandMetadata metadata =
-        FollowUpCommandMetadata.of(b -> b.batchOperationReference(batchOperation.getKey()));
-
-    // we first ask the taskResultBuilder if we even can append the all the records. We also check
-    // for adding an EXECUTE and a CONTINUE_INITIALIZATION record
-    final List<UnifiedRecordValue> sizeCheckRecords = new ArrayList<>(chunkRecords);
-    sizeCheckRecords.add(EMPTY_EXECUTION_RECORD);
-    sizeCheckRecords.add(EMPTY_INITIALIZATION_RECORD);
-    final var canAppend = taskResultBuilder.canAppendRecords(sizeCheckRecords, metadata);
-
-    if (canAppend) {
-      chunkRecords.forEach(
-          command -> {
-            LOG.trace(
-                "Appending batch operation {} chunk with {} items.",
-                batchOperation.getKey(),
-                command.getItems().size());
-            taskResultBuilder.appendCommandRecord(
-                batchOperation.getKey(), BatchOperationChunkIntent.CREATE, command, metadata);
-            metrics.recordChunkCreated(batchOperation.getBatchOperationType());
-          });
-      return true;
-    } else {
-      return false;
-    }
-  }
-
-  private static BatchOperationChunkRecord createChunkRecord(
-      final PersistedBatchOperation batchOperation, final List<Item> chunkItems) {
-    final var command = new BatchOperationChunkRecord();
-    command.setBatchOperationKey(batchOperation.getKey());
-    command.setItems(
-        chunkItems.stream().map(BatchOperationExecutionScheduler::map).collect(Collectors.toSet()));
-    return command;
-  }
-
-  private static BatchOperationItem map(final Item i) {
-    return new BatchOperationItem()
-        .setItemKey(i.itemKey())
-        .setProcessInstanceKey(i.processInstanceKey());
-  }
-
-  private void continueInitialization(
-      final TaskResultBuilder taskResultBuilder, final InitLoopState state) {
-    final var batchOperation = state.batchOperation;
-    final var command =
-        new BatchOperationInitializationRecord()
-            .setBatchOperationKey(batchOperation.getKey())
-            .setSearchResultCursor( // no null values allowed in the protocol
-                state.lastSearchResultCursor == null
-                    ? StringValue.EMPTY_STRING
-                    : state.lastSearchResultCursor)
-            .setSearchQueryPageSize(state.searchResultPageSize);
-    LOG.trace("Appending batch operation {} initializing command", batchOperation.getKey());
-    initializing.set(
-        new ExecutionLoopState(batchOperation.getKey(), command.getSearchResultCursor(), 0));
-    taskResultBuilder.appendCommandRecord(
-        batchOperation.getKey(),
-        BatchOperationIntent.INITIALIZE,
-        command,
-        FollowUpCommandMetadata.of(b -> b.batchOperationReference(batchOperation.getKey())));
-  }
-
-  private void appendFailedCommand(
-      final TaskResultBuilder taskResultBuilder,
-      final PersistedBatchOperation batchOperation,
-      final String message,
-      final BatchOperationErrorType errorType) {
-    final var batchOperationKey = batchOperation.getKey();
-    final var command = new BatchOperationPartitionLifecycleRecord();
-    command.setBatchOperationKey(batchOperationKey);
-
-    final var error = new BatchOperationError();
-    error.setType(errorType);
-    error.setPartitionId(partitionId);
-    error.setMessage(message);
-    command.setError(error);
-
-    LOG.trace("Appending batch operation {} failed event", batchOperationKey);
-    taskResultBuilder.appendCommandRecord(
-        batchOperationKey,
-        BatchOperationIntent.FAIL,
-        command,
-        FollowUpCommandMetadata.of(b -> b.batchOperationReference(batchOperationKey)));
-  }
-
-  private boolean appendExecution(
-      final Long batchOperationKey, final TaskResultBuilder taskResultBuilder) {
-    final var command = new BatchOperationExecutionRecord();
-    command.setBatchOperationKey(batchOperationKey);
-
-    LOG.trace("Appending batch operation execution {}", batchOperationKey);
-    return taskResultBuilder.appendCommandRecord(
-        batchOperationKey,
-        BatchOperationExecutionIntent.EXECUTE,
-        command,
-        FollowUpCommandMetadata.of(b -> b.batchOperationReference(batchOperationKey)));
-  }
-
-  private record ExecutionLoopState(
-      long batchOperationKey, String searchResultCursor, int numRetries) {}
-
-  /**
-   * This is a mutable state class that is used to track the state of the initialization loop. Using
-   * this state class avoids the need to pass around multiple parameters.
-   */
-  private class InitLoopState {
-    public PersistedBatchOperation batchOperation;
-    public String lastSearchResultCursor;
-    public int searchResultPageSize;
-    public ItemPage page;
-    public int keysAdded;
-    public boolean chunksAppendedThisRun = false;
-
-    public InitLoopState(final PersistedBatchOperation batchOperation) {
-      this.batchOperation = batchOperation;
-      lastSearchResultCursor = batchOperation.getInitializationSearchCursor();
-      searchResultPageSize = batchOperation.getInitializationSearchQueryPageSize(queryPageSize);
-      page = null;
-      keysAdded = 0;
-      chunksAppendedThisRun = false;
-
-      // StringValue cannot be null, so we need to check for empty string and interpret it as NULL
-      if (lastSearchResultCursor != null && lastSearchResultCursor.isEmpty()) {
-        // If the cursor is empty, we need to initialize it with the first page
-        lastSearchResultCursor = null;
-      }
-    }
-
-    public boolean hasNextPage() {
-      return page == null || !page.isLastPage();
+  record ExecutionLoopState(long batchOperationKey, String searchResultCursor, int numAttempts) {
+    public ExecutionLoopState(final PersistedBatchOperation batchOperation) {
+      this(batchOperation.getKey(), batchOperation.getInitializationSearchCursor(), 0);
     }
   }
 }
