+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/batchoperation/BatchOperationInitializer.java
@@ -1,0 +1,351 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+package io.camunda.zeebe.engine.processing.batchoperation;
+
+import static com.google.common.base.Strings.emptyToNull;
+import static com.google.common.base.Strings.nullToEmpty;
+
+import com.google.common.collect.Lists;
+import io.camunda.search.exception.CamundaSearchException;
+import io.camunda.zeebe.engine.EngineConfiguration;
+import io.camunda.zeebe.engine.metrics.BatchOperationMetrics;
+import io.camunda.zeebe.engine.processing.batchoperation.itemprovider.ItemProvider.Item;
+import io.camunda.zeebe.engine.processing.batchoperation.itemprovider.ItemProvider.ItemPage;
+import io.camunda.zeebe.engine.processing.batchoperation.itemprovider.ItemProviderFactory;
+import io.camunda.zeebe.engine.state.batchoperation.PersistedBatchOperation;
+import io.camunda.zeebe.msgpack.value.StringValue;
+import io.camunda.zeebe.protocol.impl.record.UnifiedRecordValue;
+import io.camunda.zeebe.protocol.impl.record.value.batchoperation.BatchOperationChunkRecord;
+import io.camunda.zeebe.protocol.impl.record.value.batchoperation.BatchOperationError;
+import io.camunda.zeebe.protocol.impl.record.value.batchoperation.BatchOperationExecutionRecord;
+import io.camunda.zeebe.protocol.impl.record.value.batchoperation.BatchOperationInitializationRecord;
+import io.camunda.zeebe.protocol.impl.record.value.batchoperation.BatchOperationItem;
+import io.camunda.zeebe.protocol.impl.record.value.batchoperation.BatchOperationPartitionLifecycleRecord;
+import io.camunda.zeebe.protocol.record.intent.BatchOperationChunkIntent;
+import io.camunda.zeebe.protocol.record.intent.BatchOperationExecutionIntent;
+import io.camunda.zeebe.protocol.record.intent.BatchOperationIntent;
+import io.camunda.zeebe.protocol.record.value.BatchOperationErrorType;
+import io.camunda.zeebe.stream.api.FollowUpCommandMetadata;
+import io.camunda.zeebe.stream.api.scheduling.TaskResultBuilder;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
+import org.apache.commons.lang3.RandomStringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * The is class is a scheduler that periodically checks for newly created batch operations and
+ * initializes them with the itemKeys to be executed.
+ *
+ * <p>For this, it deserializes the filter object and queries the EntityKeyProvider for all matching
+ * itemKeys for this partition. Then this collection of itemKeys will be split into smaller chunks
+ * and appended to the TaskResultBuilder as BatchOperationChunkRecord.
+ */
+public class BatchOperationInitializer {
+
+  public static final String ERROR_MSG_FAILED_FIRST_CHUNK_APPEND =
+      "Unable to append first chunk of batch operation items. Number of items: %d";
+  private static final Logger LOG = LoggerFactory.getLogger(BatchOperationInitializer.class);
+  private static final UnifiedRecordValue EMPTY_EXECUTION_RECORD =
+      new BatchOperationExecutionRecord().setBatchOperationKey(-1L);
+  private static final UnifiedRecordValue EMPTY_INITIALIZATION_RECORD =
+      new BatchOperationInitializationRecord()
+          .setBatchOperationKey(-1L)
+          .setSearchQueryPageSize(0) // random int
+          // random cursor string. 1024 chars should be enough for any cursor without sorting
+          .setSearchResultCursor(RandomStringUtils.insecure().next(1024));
+  private final int chunkSize;
+  private final int queryPageSize;
+
+  private final ItemProviderFactory itemProviderFactory;
+  private final BatchOperationMetrics metrics;
+  private final int partitionId;
+
+  public BatchOperationInitializer(
+      final ItemProviderFactory itemProviderFactory,
+      final EngineConfiguration engineConfiguration,
+      final int partitionId,
+      final BatchOperationMetrics metrics) {
+    this.itemProviderFactory = itemProviderFactory;
+    chunkSize = engineConfiguration.getBatchOperationChunkSize();
+    queryPageSize = engineConfiguration.getBatchOperationQueryPageSize();
+    this.metrics = metrics;
+    this.partitionId = partitionId;
+  }
+
+  public BatchOperationInitializationResult initializeBatchOperation(
+      final PersistedBatchOperation batchOperation, final TaskResultBuilder taskResultBuilder) {
+    if (batchOperation.isSuspended()) {
+      LOG.trace("Batch operation {} is suspended.", batchOperation.getKey());
+      return new BatchOperationInitializationResult(
+          batchOperation.getKey(), batchOperation.getInitializationSearchCursor());
+    }
+
+    final var itemProvider = itemProviderFactory.fromBatchOperation(batchOperation);
+
+    // use overall state variable for all parameters that are used in the loop
+    final InitLoopState loopState = new InitLoopState(batchOperation);
+
+    while (loopState.hasNextPage()) {
+      try {
+        loopState.page =
+            itemProvider.fetchItemPage(
+                loopState.lastSearchResultCursor, loopState.searchResultPageSize);
+      } catch (final Exception e) {
+        // in case we already have appended records to the taskResultBuilder this run, we need to
+        // save the last working cursor to the state
+        if (loopState.chunksAppendedThisRun) {
+          continueInitialization(taskResultBuilder, loopState);
+        }
+        final var exception =
+            e instanceof CamundaSearchException ? e : new CamundaSearchException(e.getMessage());
+        return new BatchOperationInitializationResult(
+            batchOperation.getKey(), loopState.lastSearchResultCursor, exception);
+      }
+
+      // Then try to append the items to the batch operation
+      final boolean appendedChunks =
+          appendChunks(batchOperation, taskResultBuilder, loopState.page.items());
+      if (appendedChunks) {
+        // everything went normally, so we can continue with the next page in the next loop
+
+        loopState.chunksAppendedThisRun = true;
+        loopState.lastSearchResultCursor = loopState.page.endCursor();
+        loopState.keysAdded += loopState.page.items().size();
+        if (loopState.page.isLastPage()) {
+          // If we have reached the last page, we can finalize the initialization and start the BO
+          // we always append the EXECUTE command at the end, even if no items were found, so we can
+          // leave the completion logic in that processor.
+          finishInitialization(loopState.batchOperation, taskResultBuilder);
+          startExecutionPhase(taskResultBuilder, loopState);
+
+          // this is the end of the initialization run, so we can break out of the loop
+
+          return new BatchOperationInitializationResult(batchOperation.getKey(), "finished");
+        }
+      } else {
+        handleFailedChunkAppend(taskResultBuilder, loopState);
+        break;
+      }
+    }
+
+    return new BatchOperationInitializationResult(
+        batchOperation.getKey(), nullToEmpty(loopState.lastSearchResultCursor));
+  }
+
+  private void startExecutionPhase(
+      final TaskResultBuilder resultBuilder, final InitLoopState state) {
+    appendExecution(state.batchOperation.getKey(), resultBuilder);
+
+    // start some metrics for the execution phase
+    metrics.recordItemsPerPartition(
+        state.batchOperation.getNumTotalItems() + state.keysAdded,
+        state.batchOperation.getBatchOperationType());
+    metrics.startStartExecuteLatencyMeasure(
+        state.batchOperation.getKey(), state.batchOperation.getBatchOperationType());
+    metrics.startTotalExecutionLatencyMeasure(
+        state.batchOperation.getKey(), state.batchOperation.getBatchOperationType());
+  }
+
+  private void finishInitialization(
+      final PersistedBatchOperation batchOperation, final TaskResultBuilder resultBuilder) {
+    final long batchOperationKey = batchOperation.getKey();
+    final var command =
+        new BatchOperationInitializationRecord().setBatchOperationKey(batchOperationKey);
+    LOG.trace("Appending batch operation {} initializing finished command", batchOperationKey);
+
+    resultBuilder.appendCommandRecord(
+        batchOperationKey,
+        BatchOperationIntent.FINISH_INITIALIZATION,
+        command,
+        FollowUpCommandMetadata.of(b -> b.batchOperationReference(batchOperationKey)));
+    metrics.recordInitialized(batchOperation.getBatchOperationType());
+  }
+
+  private void handleFailedChunkAppend(
+      final TaskResultBuilder taskResultBuilder, final InitLoopState state) {
+    if (!state.chunksAppendedThisRun) {
+      // the first chunk of this init-run could not be appended. We try to reduce the page size
+      // and retry again. If the pageSize can be halved, we will fail this partition.
+      if (state.searchResultPageSize > 1) {
+        state.searchResultPageSize = state.searchResultPageSize / 2;
+        continueInitialization(taskResultBuilder, state);
+      } else {
+        // If we failed to append the first chunk even when the pageSize is 1,
+        // we need to fail the batch operation. Otherwise, we would be stuck in an infinite loop
+        appendFailedCommand(
+            taskResultBuilder,
+            state.batchOperation,
+            String.format(ERROR_MSG_FAILED_FIRST_CHUNK_APPEND, state.page.items().size()),
+            BatchOperationErrorType.RESULT_BUFFER_SIZE_EXCEEDED);
+      }
+    } else {
+      // The RecordBatch is full, so we need another init run to continue
+      continueInitialization(taskResultBuilder, state);
+    }
+  }
+
+  /**
+   * THis method will create chunk records from the given items and append them to the
+   * taskResultBuilder. This method is atomic. Either all chunks are appended or none.
+   *
+   * @param batchOperation the batch operation to which the chunks belong
+   * @param taskResultBuilder the task result builder to append the chunks to
+   * @param items the items to be chunked and appended
+   * @return true if the chunks were successfully appended, false otherwise
+   */
+  private boolean appendChunks(
+      final PersistedBatchOperation batchOperation,
+      final TaskResultBuilder taskResultBuilder,
+      final List<Item> items) {
+
+    final var chunkRecords =
+        Lists.partition(items, chunkSize).stream()
+            .map(chunkItems -> createChunkRecord(batchOperation, chunkItems))
+            .toList();
+
+    final FollowUpCommandMetadata metadata =
+        FollowUpCommandMetadata.of(b -> b.batchOperationReference(batchOperation.getKey()));
+
+    // we first ask the taskResultBuilder if we even can append the all the records. We also check
+    // for adding an EXECUTE and a CONTINUE_INITIALIZATION record
+    final List<UnifiedRecordValue> sizeCheckRecords = new ArrayList<>(chunkRecords);
+    sizeCheckRecords.add(EMPTY_EXECUTION_RECORD);
+    sizeCheckRecords.add(EMPTY_INITIALIZATION_RECORD);
+    final var canAppend = taskResultBuilder.canAppendRecords(sizeCheckRecords, metadata);
+
+    if (canAppend) {
+      chunkRecords.forEach(
+          command -> {
+            LOG.trace(
+                "Appending batch operation {} chunk with {} items.",
+                batchOperation.getKey(),
+                command.getItems().size());
+            taskResultBuilder.appendCommandRecord(
+                batchOperation.getKey(), BatchOperationChunkIntent.CREATE, command, metadata);
+            metrics.recordChunkCreated(batchOperation.getBatchOperationType());
+          });
+      return true;
+    } else {
+      return false;
+    }
+  }
+
+  private static BatchOperationChunkRecord createChunkRecord(
+      final PersistedBatchOperation batchOperation, final List<Item> chunkItems) {
+    final var command = new BatchOperationChunkRecord();
+    command.setBatchOperationKey(batchOperation.getKey());
+    command.setItems(
+        chunkItems.stream().map(BatchOperationInitializer::map).collect(Collectors.toSet()));
+    return command;
+  }
+
+  private static BatchOperationItem map(final Item i) {
+    return new BatchOperationItem()
+        .setItemKey(i.itemKey())
+        .setProcessInstanceKey(i.processInstanceKey());
+  }
+
+  private void continueInitialization(
+      final TaskResultBuilder taskResultBuilder, final InitLoopState state) {
+    final var batchOperation = state.batchOperation;
+    final var command =
+        new BatchOperationInitializationRecord()
+            .setBatchOperationKey(batchOperation.getKey())
+            .setSearchResultCursor( // no null values allowed in the protocol
+                state.lastSearchResultCursor == null
+                    ? StringValue.EMPTY_STRING
+                    : state.lastSearchResultCursor)
+            .setSearchQueryPageSize(state.searchResultPageSize);
+    LOG.trace("Appending batch operation {} initializing command", batchOperation.getKey());
+    taskResultBuilder.appendCommandRecord(
+        batchOperation.getKey(),
+        BatchOperationIntent.INITIALIZE,
+        command,
+        FollowUpCommandMetadata.of(b -> b.batchOperationReference(batchOperation.getKey())));
+  }
+
+  void appendFailedCommand(
+      final TaskResultBuilder taskResultBuilder,
+      final PersistedBatchOperation batchOperation,
+      final String message,
+      final BatchOperationErrorType errorType) {
+    final var batchOperationKey = batchOperation.getKey();
+    final var command = new BatchOperationPartitionLifecycleRecord();
+    command.setBatchOperationKey(batchOperationKey);
+
+    final var error = new BatchOperationError();
+    error.setType(errorType);
+    error.setPartitionId(partitionId);
+    error.setMessage(message);
+    command.setError(error);
+
+    LOG.trace("Appending batch operation {} failed event", batchOperationKey);
+    taskResultBuilder.appendCommandRecord(
+        batchOperationKey,
+        BatchOperationIntent.FAIL,
+        command,
+        FollowUpCommandMetadata.of(b -> b.batchOperationReference(batchOperationKey)));
+  }
+
+  private void appendExecution(
+      final Long batchOperationKey, final TaskResultBuilder taskResultBuilder) {
+    final var command = new BatchOperationExecutionRecord();
+    command.setBatchOperationKey(batchOperationKey);
+
+    LOG.trace("Appending batch operation execution {}", batchOperationKey);
+    taskResultBuilder.appendCommandRecord(
+        batchOperationKey,
+        BatchOperationExecutionIntent.EXECUTE,
+        command,
+        FollowUpCommandMetadata.of(b -> b.batchOperationReference(batchOperationKey)));
+  }
+
+  public record BatchOperationInitializationResult(
+      long batchOperationKey, String searchResultCursor, Exception exception) {
+    public BatchOperationInitializationResult(
+        final long batchOperationKey, final String searchResultCursor) {
+      this(batchOperationKey, searchResultCursor, null);
+    }
+
+    public boolean isSuccess() {
+      return exception == null;
+    }
+  }
+
+  /**
+   * This is a mutable state class that is used to track the state of the initialization loop. Using
+   * this state class avoids the need to pass around multiple parameters.
+   */
+  private class InitLoopState {
+    public PersistedBatchOperation batchOperation;
+    public String lastSearchResultCursor;
+    public int searchResultPageSize;
+    public ItemPage page;
+    public int keysAdded;
+    public boolean chunksAppendedThisRun = false;
+
+    public InitLoopState(final PersistedBatchOperation batchOperation) {
+      this.batchOperation = batchOperation;
+      lastSearchResultCursor = batchOperation.getInitializationSearchCursor();
+      searchResultPageSize = batchOperation.getInitializationSearchQueryPageSize(queryPageSize);
+      page = null;
+      keysAdded = 0;
+      chunksAppendedThisRun = false;
+
+      // StringValue cannot be null, so we need to check for empty string and interpret it as NULL
+      // If the cursor is empty, we need to initialize it with the first page
+      lastSearchResultCursor = emptyToNull(batchOperation.getInitializationSearchCursor());
+    }
+
+    public boolean hasNextPage() {
+      return page == null || !page.isLastPage();
+    }
+  }
+}
